mobile changes and some other stuff
