+++ ui/client/src/components/LoginPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { IDarkThemeProps } from "../App";
-import Header from "./Header";
+import MobileHeader from "./MobileHeader";
 import { useMobile } from "./helpers/CustomHooks";
 import { Classes, Card, H1, FormGroup, InputGroup, Button, Intent, Callout, H4 } from "@blueprintjs/core";
 import { useTranslation } from "react-i18next";
@@ -128,7 +128,7 @@ export default function LoginPage(props: IDarkThemeProps & { setAuthenticated: (
   );
 
   return <>
-    <Header
+    <MobileHeader
       darkThemeProps={props}
       className='login-header'
     />
