+++ ui/client/src/components/MobileHeader.scss
@@ -8,50 +8,24 @@
   }
   position: relative;
 }
-.header-wrapper {
-  height: 85px;
-
-  .desktop-header {
-    max-width: 1280px;
-    margin: auto;
-    margin-top: 5px;
-    margin-bottom: 10px;
-    padding: 0 20px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    max-height: 75px;
-
-    a {
-      display: flex;
-      align-items: center;
-      .App-logo {
-        max-height: 75px;
-      }
-    }
-
-    .settings {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-around;
-      align-items: center;
-      height: 75px;
-      padding: 5px;
-    }
-  }
-}
-.#{$bp-ns}-dark {
-  .header-wrapper, .mobile-header-content {
-    box-shadow: unset;
-  }
-}
 
 .mobile-header-wrapper {
   .mobile-header-content {
+    height: $pt-navbar-height;
+    padding: 0 15px;
+    position: fixed;
+    top: 0;
+    right: 0;
+    left: 0;
+    width: 100%;
+    z-index: 10;
     display: flex;
     flex-direction: row;
     justify-content: space-between;
+    background-color: $white;
+    .#{$bp-ns}-dark & {
+      background-color: $dark-gray5;
+    }
 
     .left-align {
       display: flex;
@@ -87,12 +61,16 @@
     position: fixed;
     top: 50px;
     left: 0;
-    width: calc(100% - 0px);
+    width: 100%;
     height: auto;
     padding: unset;
     border-top-left-radius: 0;
     border-top-right-radius: 0;
     z-index: 10;
+    .#{$bp-ns}-dark & {
+      background-color: $dark-gray5;
+      box-shadow: 0 0 0 1px rgba(17,20,24,.2),0 0 0 rgba(17,20,24,0),0 1px 1px rgba(17,20,24,.4);
+    }
     .settings {
       margin: 0 20px;
       padding: 10px 0;
