+++ ui/client/src/components/recipeList/RecipeListItem.scss
@@ -132,7 +132,10 @@
         font-style: italic;
         text-align: end;
         margin: unset;
-        color: $pt-dark-link-color;
+        color: $pt-link-color;
+        .#{$bp-ns}-dark & {
+          color: $pt-dark-link-color;
+        }
       }
     }
   }
