+++ ui/client/src/static/locales/de/translation.json
@@ -93,8 +93,8 @@
   "commentNotModified": "Der Kommentar wurde nicht verändert.",
   "commentNotSaved": "Beim Speichern des Kommentars ist ein Fehler aufgetreten.",
   "commentSaved": "Der Kommentar wurde erfolgreich gespeichert.",
-  "comments": "Ein Kommentar",
   "comments_0": "Sei der erste der kommentiert:",
+  "comments_1": "Ein Kommentar",
   "comments_plural": "{{count}} Kommentare",
   "editComment": "Kommentar bearbeiten...",
   "newComment": "neuer Kommentar",
