+++ ui/client/src/static/locales/en/translation.json
@@ -91,9 +91,9 @@
   "save": "Save",
   "commentSaved": "The comment was successfully saved.",
   "commentNotModified": "The comment has not been modified.",
-  "comments": "One Comment",
   "comments_0": "Be the first to comment:",
-  "comments_plural": "{{count}} Comments",
+  "comments_1": "One comment",
+  "comments_plural": "{{count}} comments",
   "editComment": "Edit the comment...",
   "addComment": "Add a comment to the recipe...",
   "comment": "Comment",
